+++ starter_code/models/movie.js
@@ -1,0 +1,15 @@
+const mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+const movieSchema = new Schema({
+  title: String,
+  director: String,
+  stars: Array,
+  image: String,
+  description: String,
+  showtimes: Array,
+});
+
+const Movie = mongoose.model('Movie', movieSchema);
+
+module.exports = Movie;
