+++ starter_code/routes/movies.js
@@ -1,10 +1,28 @@
 const express = require('express');
-const movies = require('Movie');
+const Movie = require('../models/movie.js');
 const router = express.Router();
 
 /* GET users listing. */
 router.get('/', function(req, res, next) {
-  res.render('movies', movies);
+  Movie.find()
+  .then(movies =>{
+    res.render('movies', {movies});
+  })
+  .catch(error =>{
+    console.log('upsi: ', error)
+  }) 
 });
+
+router.get('/:id', (req, res, next) => {
+  const id = req.params.id
+  Movie.findById(id)
+  .then(movie => {
+    res.render('movie', { movie: movie });
+  }) 
+  .catch(error => {
+    console.log(error);
+  })
+});
+
 
 module.exports = router;
